Adds fixed and design domain blocks to Mesh

+++ functional/plato/mesh/MeshProxy.hpp
@@ -22,7 +22,7 @@ struct Density
 /// and a density field.
 struct MeshProxy
 {
-    using BlockIDType = int64_t;
+    using BlockIDType = unsigned;
     using DensityVector = std::vector<Density>;
     using BlockDensities = std::map<BlockIDType, DensityVector>;
